Linter complaisance and reorganization
